Add 'the-birthday-bar' challenge test cases and solution
This commit adds a new challenge called 'the-birthday-bar' to our testing suite. It includes the main solution file, multiple test case input and output files, as well as the test script itself. These enhancements will facilitate the verification of the correctness and robustness of the implemented solution by facilitating automatic testing with various inputs and expected outputs.