Add 'Day of the Programmer' challenge
This commit added the 'Day of the Programmer' challenge. Files include main code and test files, sample and test input/output files. The challenge is about calculating the date of the 256th day of a given year, considering leap years and the transition in calendar year 1918, which makes the problem tricky. The solution uses calendar related calculations and conditions to get the correct date. Tests are included to validate the implemented functionality.