Add 'drawing-book' challenge with its solutions
This commit introduces a new challenge called 'drawing-book'. New Python script 'main.py' was created to provide the solution for this challenge. It also includes the corresponding 'input' and 'output' directories which contain the input data and expected output results for the challenge respectively. In addition, a test script 'test.py' was added to run unit tests to verify the implementation of the solution.